Removed datetime as trait, moved it to plugin updated tests
